+++ persistent-mysql/test/InsertDuplicateUpdate.hs
@@ -61,12 +61,15 @@ specs = describe "DuplicateKeyUpdate" $ do
       dbItems <- map entityVal <$> selectList [] []
       sort dbItems @== sort (newItem : items)
     it "updates existing records" $ db $ do
+      let postUpdate = map (\i -> i { itemQuantity = fmap (+1) (itemQuantity i) }) items
       deleteWhere ([] :: [Filter Item])
       insertMany_ items
       insertManyOnDuplicateKeyUpdate
         items
         []
         [ItemQuantity +=. Just 1]
+      dbItems <- sort . fmap entityVal <$> selectList [] []
+      dbItems @== sort postUpdate
     it "only copies passing values" $ db $ do
       deleteWhere ([] :: [Filter Item])
       insertMany_ items