+++ persistent-redis/ChangeLog.md
@@ -1,3 +1,8 @@
+## 2.12.0.1 (unreleased)
+
+* [#1123](https://github.com/yesodweb/persistent/pull/1223):
+    * Changed the error message from trying to serialize a `PersistDbSpecific` value into `PersistLiteral_`.
+
 ## 2.12.0.0
 
 * Decomposed `HaskellName` into `ConstraintNameHS`, `EntityNameHS`, `FieldNameHS`. Decomposed `DBName` into `ConstraintNameDB`, `EntityNameDB`, `FieldNameDB` respectively. [#1174](https://github.com/yesodweb/persistent/pull/1174)